+++ src/components/index.tsx
@@ -11,6 +11,7 @@ import TwToast from './tw-toast';
 import TwTable from './tw-table';
 import TwSearch from './tw-search';
 import TwSelect from './tw-select';
+import TwListbox from './tw-listbox';
 
 export {
   AppLogo,
@@ -26,4 +27,5 @@ export {
   TwTable,
   TwSearch,
   TwSelect,
+  TwListbox
 };
